% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{row_sums_rust}
\alias{row_sums_rust}
\title{Compute row sums of a matrix using faer}
\usage{
row_sums_rust(matrix)
}
\arguments{
\item{matrix}{A numeric matrix}
}
\value{
A numeric vector of row sums
}
\description{
Compute row sums of a matrix using faer
}
